Add check for existing data to prevent data loss

+++ rip-cd.sh
@@ -1,20 +1,38 @@
 #!/bin/bash
 
-if [ $# -eq 0 ]; then
 DEVICE="/dev/cdrom"
-else
-DEVICE="$1"
+CDDB="-L 0 --cddbp-server=gnudb.gnudb.org"
+
+while [ $# -gt 0 ]
+do
+    case "$1" in
+        --no-cddb)
+            CDDB=""
+            ;;
+        /dev/*)
+            DEVICE=$1
+            ;;
+        *)
+            echo "Option $1 is not detected. Aborting."
+            exit 1
+            ;;
+    esac
+    shift
+done
+
+echo "Using device $DEVICE for ripping."
+
+if [ -e tmp ]
+then
+    echo 'The folder tmp seems to be existing. To avoid data loss, this script terminates here.'
+    exit 1
 fi
 
 mkdir -p tmp
 cd tmp
 
 rm -rf *
-cdda2wav -L 0 -B -D "$DEVICE"
-
-#cp audio.cddb ..
-
-#normalize-audio -b audio_*.wav
+cdda2wav $CDDB -B -D "$DEVICE"
 
 rm -rf titles.info album.info
 
@@ -22,4 +40,4 @@ cd ..
 
 eject "$DEVICE"
 
-echo "Bitte nun prepare-rip.sh aufrufen, um die Track-Informationen zu extrahieren."
+echo "Please call prepare-rip.sh now to extarct the track information from the rip."
